requirements.txt
click==6.6
pyyaml==3.11
behave==1.2.5
nose==1.3.4
gitpython==2.1.1
Flask==0.12.0
requests==2.12.4
RandomWords==0.1.5
boto3==1.4.7
pymongo==3.6.0
